% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_bootstrap_sampling-v0.1.R
\name{boot.sample}
\alias{boot.sample}
\title{Cluster Boostrap sampling}
\usage{
boot.sample(dataset, column_obs, column_group)
}
\arguments{
\item{dataset}{is the dataset to be resampled}

\item{column_obs}{is the column of the dataset containing the obsevations. must be in the format dataset$observation}

\item{column_group}{is the column of the dataset containing the groups (cluster). must be in the format dataset$group}
}
\description{
This function resample a dataset where observation are grouped by one of its variables. Internally used by \link{spt_glmmPQL_bts_val}
}
\details{
The function is a modified version of the cluster bootstrap scheme coded by Bouwmeester et al 2012
and available on-line at  http://aje.oxfordjournals.org/content/suppl/2013/05/09/kws396.DC1/kws396_Web_Material.pdf

This modified version is only producing the resampled dataset.
}
\references{
Bouwmeester, W., K. G. M. Moons, T. H. Kappen, W. A. van Klei, J. W. R. Twisk, M. J. C. Eijkemans, and Y. Vergouwe. 2013. Internal Validation of Risk Models in Clustered Data: A Comparison of Bootstrap Schemes. American Journal of Epidemiology 177:1209–1217.
}
\seealso{
\link{spt_glmmPQL_bts_val}
}
\author{
Filippo Ferrario, \email{filippo.f3rrario.1@gmail.com}
}
