% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density2list.R
\name{density2list}
\alias{density2list}
\title{Apply density.ppp to element of a list of ppp}
\usage{
density2list(
  to_density = NULL,
  W_from = NULL,
  full_list = W_from,
  changeW = FALSE,
  addZeros = TRUE,
  ...
)
}
\arguments{
\item{to_density}{a solist of ppp for which density needs to be computed}

\item{W_from}{a list windows (objects of class "owin") or data acceptable to \link{as.owin}, where to look for the Window of observation to assign to to_density elements}

\item{full_list}{optional. A character vector or a list of spatstat objects (e.g., owin or ppp) defining the complete set of replicates for which a variable need to be computed. To be specified if "full_list" is a subset of W_from to avoid superflouos calculations.}

\item{changeW}{logical (default =FALSE). If TRUE  and "W_from" is specified, the Window of the elements from "W_from" is used and will replace that of elements in "to_density" (this in case X is a ppp). If FALSE the window of "to_density" elements is kept.}

\item{addZeros}{logical. If TRUE (the default) and "W_from" is specified, the output will include elements with all pixels having value 0 when an element is in "W_from" but not in "to_density"}

\item{...}{to pass extra arguments to \link{density.ppp}}
}
\value{
an object of class
}
\description{
This function is to be used to create a variable (usually a predictor) using \link{density.ppp} for each replicate plot to be analyzed.
The intended use cases are:
\itemize{
\item a set of ppp is to be used as response variable
\item for each ppp, the analyst wants to produce a variable describing a given feature using \link{density.ppp}, in replicate plots where another set of spatial objects (e.g., owin, ppp) has been observed.
However it is possible that the predictor ppp is not observed (i.e., a real zero and not to be considered missing data) in all the plots where the response ppp is available.
In this case the function produce an image whose pixels have all 0 value
}
}
\details{
List "to_density" may be a subset of or longer list than "full_list" or "W_from".
Only element from "to_density" also present in either "W_from" or "full_list" will be considered.

\strong{If addZeros=TRUE}
For those element of "full_list" not in "to_density", the function creates "0": the absence elements of list "full_list" from the "to_density" list is not to be considered as missing values but rather as 0.
For those element the function can create an im object in which the pixels value is set to zero.

The 'full_list' may be a subset of "W_from" in case not all the elements in "W_from" need to be used for the analysis.
Lists are compared by names so technically the order should not be important.

If some elements in "to_density" are not contained in "full_list" (either not present or misspelled) a warning is given.
If "full_list" is specified, all its elements (either characters or names of a solist) must be included in "W_from" otherwise an error is thrown.
}
\seealso{
\link{density.ppp}, \link{densityfun}
}
\author{
Filippo Ferrario, \email{filippo.f3rrario@gmail.com}
}
