% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/focal2list.R
\name{focal2list}
\alias{focal2list}
\title{Apply spstFocal to the elements of the list "to_focal" with potential introduction of real zeros.}
\usage{
focal2list(
  to_focal = NULL,
  W_from = NULL,
  full_list = W_from,
  changeW = TRUE,
  addZeros = TRUE,
  ...
)
}
\arguments{
\item{to_focal}{a list of spatstat objects (e.g., owin or ppp) to which apply the focal function.}

\item{W_from}{optional if to_focal is a list of ppp. A list of windows (objects of class "owin") or data acceptable to \link[spatstat.geom:as.owin]{spatstat.geom::as.owin}, where to look for the Window of observation to assign to to_focal elements.}

\item{full_list}{optional. A character vector or a list of spatstat objects (e.g., owin or ppp) defining the complete set of replicates for which a variable need to be computed using \link{spstFocal}. To be specified if "full_list" is a subset of "W_from" to avoid superfluous calculations. If not specified but "W_from" is, then "full_list" is set to be "W_from".}

\item{changeW}{logical. If TRUE (the default) and "W_from" is specified, the Window of the elements from "W_from" is used and will replace that of elements in "to_focal" (this in case X is a ppp). If FALSE the window of "to_focal" elements is kept.}

\item{addZeros}{logical. If TRUE (the default) and "W_from" is specified, the output will include elements with all pixels having value 0 when an element is in "W_from" but not in "to_focal"}

\item{...}{agruments to be passed to \link{spstFocal} and \link[raster:focal]{raster::focal} (In particular \code{fun}, \code{na.rm}, \code{pad} and \code{padValue}).}
}
\value{
an object of class "imlist"  "solist"  "anylist" "listof"  "list"
}
\description{
This function is to be used to create a variable (usually a predictor)  using \link{spstFocal} for each replicate plot to be analyzed.
The intended use cases:
\itemize{
\item a set of ppp is to be used as response variable
\item for each ppp, the analyst wants to produce a variable describing a given feature using a focal statistic, in replicate plots where another set of spatial objects (e.g., owin, ppp) has been observed.
However it is possible that that feature is not observed (i.e., a real zero and not to be considered missing data) in all the plots where the response ppp is available.
In this case the function produce an image whose pixels have all 0 value.
}
}
\details{
List "to_focal" may be a subset of or longer list than "full_list" or "W_from". Only element from "to_focal" also present in either "W_from" or "full_list" will be considered.
For those element of "full_list" not in "to_focal", the function creates "0": the absence elements of list "full_list" from the "to_focal" list is not to be considered as missing values but rather as 0.
For those element the function can create an im object in which the pixels value is set to zero.

The 'full_list' may be a subset of "W_from" in case not all the elements in "W_from" need to be used for the analysis.
Lists are compared by names so technically the order should not be important.

If some elements in "to_focal" are not contained in "full_list" (either not present or misspelled) a warning is given.
If "full_list" is specified, all its elements (either characters or names of a solist) must be included in "W_from" otherwise an error is thrown.

\emph{\strong{ATTENTION}}
Verify to check the value of the argument DivdideByPixelArea passed to \link[spatstat.geom:pixellate]{spatstat.geom::pixellate} via \link{spstFocal}.
Default is FALSE!
}
\seealso{
\link{spstFocal}, \link[spatstat.geom:pixellate]{spatstat.geom::pixellate}, \link[spatstat.geom:pixellate.owin]{spatstat.geom::pixellate.owin}, \link[spatstat.geom:pixellate.ppp]{spatstat.geom::pixellate.ppp}, \link{focal}
}
\author{
Filippo Ferrario, \email{filippo.f3rrario@gmail.com}
}
