% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_hyp.R
\name{merge_hyp}
\alias{merge_hyp}
\title{Function to merge 2 hyperframes}
\usage{
merge_hyp(
  x,
  y,
  by = intersect(names(x), names(y)),
  by.x = by,
  by.y = by,
  sort = TRUE,
  suffixes = c(".x", ".y"),
  no.dups = TRUE,
  incomparables = NULL,
  ...
)
}
\arguments{
\item{x}{an hyperframe (sensu spatstat)}

\item{y}{an hyperframe (sensu spatstat)}

\item{by, by.x, by.y}{specifications of the columns used for merging. See \link{merge}}

\item{suffixes}{as in \link{merge}}

\item{no.dups}{as in \link{merge}}

\item{incomparables}{as in \link{merge}}

\item{...}{as in \link{merge}}
}
\description{
Function to merge 2 hyperframes
}
\details{
Modified from merge.data.frame.
Currently only implemented to perform merge with all=FALSE
Appending a suffix and renaming of variables is not implemented
}
\examples{

}
\seealso{
\link{hyperframe}, \link{merge}
}
\author{
Filippo Ferrario, \email{filippo.f3rrario@gmail.com}
}
