% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met2raster.R
\name{met2raster}
\alias{met2raster}
\title{Metric to Raster}
\usage{
met2raster(obs_polygon, side.cell = 0.1, FUN, make_spatstat_im = TRUE, ...)
}
\arguments{
\item{obs_polygon}{polygon delimiting the area observed (e.g., a transect)}

\item{side.cell}{the size of the (squared) pixel side to be used to subdivide the observation window (i.e., the polygon defining the extent of the raster to be produced).}

\item{make_spatstat_im}{logical (default=TRUE). Should the output be a \link{im} object for \code{spatstat} rather than a \link[raster:raster]{raster::raster}?}

\item{...}{the distance value that will be used to calculate the Area:distance ratio when \code{cx} falls in a patch. All distances < than \code{zero_distance} will be set to this.}
}
\description{
Metric to Raster
}
\seealso{
\link{PRindex_body}, \link[sf:st]{sf::st_point}, \link[sf:geos_unary]{sf::st_buffer}
}
\author{
Filippo Ferrario, \email{filippo.f3rrario@gmail.com}
}
