% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel.cor.R
\name{panel.cor}
\alias{panel.cor}
\title{Correlation Pair Plot panel}
\usage{
panel.cor(x, y, digits = 2, prefix = "", cex.cor, cor.treshold = NULL, ...)
}
\description{
function to be produce panels in pair plots to investigate correlations
}
\details{
Modified from Zurr 2009 book.
}
\author{
Filippo Ferrario, \email{filippo.f3rrario.1@gmail.com}
}
