% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patches_stats.R
\name{patch_stats}
\alias{patch_stats}
\title{Number of patch in buffer}
\usage{
patch_stats(poly, buffer, cx, field)
}
\description{
Number of patch within
}
\seealso{
\link{map2raster}, \link{PatchRelv}, \link[sf:st]{sf::st_point}, \link[sf:geos_unary]{sf::st_buffer}
}
\author{
Filippo Ferrario, \email{filippo.f3rrario@gmail.com}
}
