% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_patches.R
\name{PatchNum}
\alias{PatchNum}
\title{Number of patch in buffer}
\usage{
PatchNum(poly, buffer, cx)
}
\description{
Number of patch in buffer
}
\seealso{
\link{map2raster}, \link{PatchRelv}, \link[sf:st]{sf::st_point}, \link[sf:geos_unary]{sf::st_buffer}
}
\author{
Filippo Ferrario, \email{filippo.f3rrario@gmail.com}
}
