% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PatchRelv.R
\name{PatchRelv}
\alias{PatchRelv}
\title{Patch Relevance index calculated within a buffer.}
\usage{
PatchRelv(poly, buffer, cx, zero_distance = 0.01)
}
\arguments{
\item{poly}{a collection of polygons representing the patches to be use to calculate the index}

\item{buffer}{a vector (or dataframe) of polygons representing the buffers. tipically the result of sf::st_buffer.}

\item{cx}{st_point center(s) of the buffer. Provided by \code{map2raster} if \code{PatchRelv} is used in it.}

\item{zero_distance}{the distance value that will be used to calculate the Area:distance ratio when \code{cx} falls in a patch. All distances < than \code{zero_distance} will be set to this.}
}
\description{
Patch Relevance index calculated within a buffer.
}
\details{
The \verb{Patch Relevance} is calculated as:
\deqn{PR = \sum{n=1}^{n} \frac{A}{D}}
where \emph{A} is the area of the \emph{i} polygon in a buffer of a given radius, and \emph{D} is the distance of that polygon to the center of the buffer.

\link{PRindex} is an alternative version, probably easier to use for tests.
}
\seealso{
\link{PRindex_body}, \link[sf:st]{sf::st_point}, \link[sf:geos_unary]{sf::st_buffer}, \link{PRindex},  \link{map2raster}
}
\author{
Filippo Ferrario, \email{filippo.f3rrario@gmail.com}
}
