% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PerimAreaRatio.R
\name{peri_area}
\alias{peri_area}
\title{Mean perimeter to area ratio of patches within a buffer.}
\usage{
peri_area(poly, buffer, ...)
}
\arguments{
\item{poly}{a collection of polygons representing the patches to be use to calculate the index}

\item{buffer}{a vector (or dataframe) of polygons representing the buffers. tipically the result of sf::st_buffer.}

\item{...}{unused.}
}
\description{
Mean perimeter to area ratio of patches within a buffer.
}
\details{
The ratio is calculated as:
\deqn{ \frac{Perimeter}{Area}}
it is not scale independent, meaning that increasing the patch size while not changing the patch form will change the ratio.

The \code{...} argument allows the functioning within \link{map2raster}, but it does not supply used arguments.
}
\seealso{
\link[sf:st]{sf::st_point}, \link[sf:geos_unary]{sf::st_buffer}, \link{PRindex}, \link{map2raster}
}
\author{
Filippo Ferrario, \email{filippo.f3rrario@gmail.com}
}
