% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_quadrat.R
\name{point_quadrat}
\alias{point_quadrat}
\title{Function to sample random points in random quadrats in a shapefile-like file.}
\usage{
point_quadrat(
  dsn = NULL,
  file_tag = NULL,
  shp = NULL,
  site_poly = NULL,
  n_quad = 3,
  n_pts = 6,
  seed = 1294,
  quadrat_size = 1,
  class_lab = "class",
  class_val = c("Bloc ou galet", "SC)diment mixte grossier")
)
}
\arguments{
\item{dsn}{path where to save output}

\item{file_tag}{otpional. A tag to be attached at the name of files to dfferentiate them from others from the same function.}

\item{shp}{polygon shapefile that need to be sampled.}

\item{site_poly}{optional. Polygon delimiting the area of the site to be taken into account. If not provided the shapefile provided in 'shp' will be consitered as site.}

\item{n_quad}{number of replicate quadrats}

\item{n_pts}{number of points per category per quadrats}

\item{seed}{seed for random generation if needed.}

\item{quadrat_size}{size of the replicare quadrat in meters.}

\item{class_lab}{Character giving the name of the column storing the classification of polygons}

\item{class_val}{Character giving the classification levels for which sampling should be performed}
}
\description{
Function to sample random points in random quadrats in a shapefile-like file.
}
\details{
Attention about the behaviour of the function if the shp shapefile is referring to multiple sites. In this case it is better to provide the parameter "site_poly"
}
\author{
Filippo Ferrario, \email{filippo.f3rrario@gmail.com}
}
