% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRindex.R
\name{PRindex}
\alias{PRindex}
\title{Patch Relevance index calculated within a buffer.}
\usage{
PRindex(cx, poly, buffer = 5, zero_distance = 0.01)
}
\arguments{
\item{cx}{an st_point center of the buffer}

\item{poly}{a collection of polygons representing the patches to be use to calculate the index}

\item{buffer}{the radius of the buffer to use around \code{cx}.}

\item{zero_distance}{the distance value that will be used to calculate the Area:distance ratio when \code{cx} falls in a patch. All distances < than \code{zero_distance} will be set to this.}
}
\description{
Patch Relevance index calculated within a buffer.
}
\details{
The \code{PRindex} is calculated as:
\deqn{PR = \sum{n=1}^{n} \frac{A}{D}}
where \emph{A} is the area of the \emph{i} polygon in a buffer of a given radius, and \emph{D} is the distance of that polygon to the center of the buffer.
}
\seealso{
\link{PRindex_body}, \link[sf:st]{sf::st_point}, \link[sf:geos_unary]{sf::st_buffer}
}
\author{
Filippo Ferrario, \email{filippo.f3rrario@gmail.com}
}
