% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch_relevance.R
\name{PRindex_body}
\alias{PRindex_body}
\title{Patch Relevance index calculated within a buffer.}
\usage{
PRindex_body(cx, poly, one_body_length = 0.03, buffer = 5)
}
\arguments{
\item{cx}{an st_point center of the buffer}

\item{poly}{a collection of polygons representing the patches to be use to calculate the index}

\item{buffer}{the radius of the buffer to use around \code{cx}.}

\item{one_body_lenght}{the lenght to be used to standardize the distance: e.g. the mean length or the radius of the body for a given species.}

\item{zero_distance}{the distance value that will be used to calculate the Area:distance ratio when \code{cx} falls in a patch. All distances < than \code{zero_distance} will be set to this.}
}
\description{
Patch Relevance index calculated within a buffer.
}
\details{
The \code{PRindex} is calculated as:
$$PR = \sum_{n=1}^{n} \frac{A}{D}$$

The \code{PRindex_body} is calculated as:
$$SR = \sum_{n=1}^{n} \frac{A}{\frac{D}{mean~~body~~lenght}}$$

where \emph{A} is the area of the \emph{i} polygon in a buffer of a given radius, and \emph{D} is the distance of that polygon to the center of the buffer.
}
\seealso{
\link[sf:st]{sf::st_point}, \link[sf:geos_unary]{sf::st_buffer}
}
\author{
Filippo Ferrario, \email{filippo.f3rrario@gmail.com}
}
