% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMSE_POISSON_mixed_pred-v0.1.R
\name{rmsePOI}
\alias{rmsePOI}
\title{Root Mean squared error for Poisson glm(m)}
\usage{
rmsePOI(model, Y_col = NULL, ...)
}
\arguments{
\item{model}{a Poisson glm(m) model}

\item{Y_col}{name of the column containing the response variable.}

\item{...}{used to pass arguments to predict.}
}
\description{
Root Mean squared error for both fixed and mixed models FITTED WITH POISSON FAMILY
}
\details{
\code{...} can be used to specify "level" . 0= To match re.form=NA in Bouwmeester et al 2013.
See \code{?predict.lme} : Level values increase from outermost to innermost grouping, with level zero corresponding to the population predictions.
}
\seealso{
\link[nlme:predict.lme]{nlme::predict.lme}
}
\author{
Filippo Ferrario, \email{filippo.f3rrario@gmail.com}
}
