% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spstFocal.R
\name{spstFocal}
\alias{spstFocal}
\title{Focal stats to spatstats}
\usage{
spstFocal(X, W_from = NULL, dist = NULL, side.cell = NULL, ...)
}
\arguments{
\item{X}{a ppp or a owin object}

\item{W_from}{a window (object of class "owin") or data acceptable to \link[spatstat:as.owin]{spatstat::as.owin}. Needed to give a spatial domain if X is an owin. Optional if X is a ppp.}

\item{dist}{the size of the side of the squared moving window used to calculate the focal statistic.}

\item{side.cell}{the size of the (squared) pixel side to be used to subdivide the moving window.}

\item{...}{agruments to be passed to pixellate (except "eps";e.g., DivideByPixelArea) and \link[raster:focal]{raster::focal} (in particular \code{fun}, \code{na.rm}, \code{pad} and \code{padValue}).}
}
\value{
an \code{im} object
}
\description{
This function to apply the raster::focal funtion to spatstat objects
}
\details{
\subsection{If X is an owin object, the "W_from" argument is used to assign the Window of observation which otherwise will be lacking.
In this case the window should likely be assigned from a ppp under analysis.
#'
\emph{\strong{workflow}}}{

If X is an \code{owin} object:
\enumerate{
\item pixellate the owin object with the "W_from" window used to assign the spatial domain.
The pixallate function is used with DivideByPixelArea=T.
As such, the vaule assigned to each pixel is the percent of the pixel coverd by the polygon being pixellated.
if "W_from" has an irregual shape the values outside "W_from" are given a value 0.
\item assign the spatial domain of the "W_from" to exclude the pixels outside "W_from".
This is achieved passing the pixellate output to as.im(,W=W_from) which set the pixels outside "W_from" as NA.
\item convert image to a rasterLayer to be able to use raster::focal
}

If X is an \code{ppp} object:
the density map is calculated using pixellate.ppp and transformed in a raster to be able to use the function focal.
If the argument "W_from" is specified then the Window of X is changed to correspond to that of "W_from". Oterwise X retain its original Window.
Setting "W_from" when X is a \code{ppp} object is useful to set the Window of the density variable equal to the that of another ppp (e.g., that used a response variable)
when X has been created with a different Window. NOTE: only the spatial domain of the density variable is modified and not the window of the \code{ppp} used as X.

Common steps:
4) apply focal to raster (the matrix need to have numebr of cells by row and column odd)
5) extract the data matrix and return the output as an \code{im} object
}
}
\examples{

}
\seealso{
\link[spatstat:pixellate.owin]{spatstat::pixellate.owin}, \link[spatstat:pixellate.ppp]{spatstat::pixellate.ppp}, \link[raster:focal]{raster::focal}
}
\author{
Filippo Ferrario, \email{filippo.f3rrario@gmail.com}
}
