% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_glm_bootstrap_validation-v0.1.R
\name{spt_glm_bts_val}
\alias{spt_glm_bts_val}
\title{Internal Validation with bootstrap}
\usage{
spt_glm_bts_val(mod_fmla = NULL, Y_col = NULL, B = 200, dataset = NULL, ...)
}
\arguments{
\item{mod_fmla}{character vectror with formulas of the models to be fit. It must be a named vector regardless of its length.}

\item{Y_col}{character name of the column with response variables}

\item{B}{number of bootstrap samples to be taken. If 0 no bootstrap is performed and the model fitted on the whole dataset returned.}

\item{dataset}{full dataset}

\item{...}{not used}
}
\value{
A dataframe from which calculate the average optimism to be used to correct the model fitted on the original dataset.
Each line is a parameter of one particular model in the model set, and the averaged one\link{TO BE IMPLEMENTED}, for each iteration
The column 'orig' state if a row refer to the original dataframe or a bootstrap sample.
}
\description{
Function to fit a set of GLM models to both whole dataset and replicate bootstapped datasets.
The function extract the model parameters, RMSE and R^2^ of each model runs, and calculate the RMSE optimism at each run.
}
\details{
The function fits the models in the model set \code{mod_fmla} to the origial dataframe and to each of the boostrapped samples.
Bootstrap sampling is performed internally.
\link{TO BE IMPLEMENTED} At each iteration the models in the model set are also averaged and the parameter of the shrinked model saved.

The R^2^ is calculated using \link[uMIn:r.squaredGLMM]{uMIn::r.squaredGLMM}
}
\author{
Filippo Ferrario, \email{filippo.f3rrario.1@gmail.com}
}
