% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_glmmPQL_bootstrap_validation-v0.1.R
\name{spt_glmmPQL_bts_val}
\alias{spt_glmmPQL_bts_val}
\title{Internal Validation with bootstrap for Spatial point process}
\usage{
spt_glmmPQL_bts_val(
  mod_fmla = NULL,
  random_fmla = NULL,
  Y_col = NULL,
  B = 200,
  dataset = NULL,
  column_group = NULL,
  RAN.pred = FALSE,
  ...
)
}
\arguments{
\item{mod_fmla}{character vectror with formulas of the models to be fit. It must be a named vector regardless of its length.}

\item{random_fmla}{string with random formula of the models to be fit}

\item{Y_col}{character name of the column with response variables}

\item{B}{number of bootstrap samples to be taken. If 0 no bootstrap is performed and the model fitted on the whole dataset returned.}

\item{dataset}{full dataset}

\item{column_group}{character name of the dataset column with grouping.}

\item{RAN.pred}{logical. Specify if using or not the random factor to obtain predictions. Defaults to FALSE. see Details and \link[nlme:predict.lme]{nlme::predict.lme}}

\item{...}{used to pass arguments to \link[MuMIn:r.squaredGLMM]{MuMIn::r.squaredGLMM} (e.g. pj2014)}
}
\value{
A dataframe where each line is a parameter of one particular model in the model set, and the averaged one (TO BE IMPLEMENTED), for each iteration
The column 'type' state if a row refer to the original dataframe or a bootstrap sample.

The \code{RAN.pred} argument determines on which predictions the RMSE is calculated (i.e. including or not the Random factor). It does not affect the R^2^ values.

TO BE IMPLEMENTED: At each iteration the models in the model set are also averaged and the parameter of the shrinked model saved.

For using Random factor when obtaining predictions please see Bouwmeester et al 2013.
}
\description{
Function to fit models in a set to both whole dataset and replicate bootstapped datasets using GLMM PQL.
The function extract the model parameters, RMSE and R^2^ (marginal and conditionals) of each model runs, and calculate the RMSE optimism at each run.
}
\details{
The function fits the models in the model set \code{mod_fmla} to the origial dataframe and to each of the boostrapped samples.
This function should be used to produce the dataframe from which calculate the average optimism and correct the model fitted on the original dataset.

Cluster Boostrap sampling is performed using \link{boot.sample}.

Models are fitted using \link[MASS:glmmPQL]{MASS::glmmPQL} with settings specified to mimic the fitting done for MIXED models by \link[spatstat:mppm]{spatstat::mppm}.
}
\references{
Bouwmeester, W., K. G. M. Moons, T. H. Kappen, W. A. van Klei, J. W. R. Twisk, M. J. C. Eijkemans, and Y. Vergouwe. 2013. Internal Validation of Risk Models in Clustered Data: A Comparison of Bootstrap Schemes. American Journal of Epidemiology 177:1209–1217.
Harrell, F. E. 2015. Regression Modeling Strategies. Springer International Publishing, Cham. (Chapter 5)
}
\seealso{
\link{boot.sample}, \link[MASS:glmmPQL]{MASS::glmmPQL}, \link[MuMIn:r.squaredGLMM]{MuMIn::r.squaredGLMM}, \link[spatstat:mppm]{spatstat::mppm}
}
\author{
Filippo Ferrario, \email{filippo.f3rrario.1@gmail.com}
}
