% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/win2ppp.R
\name{win2ppp}
\alias{win2ppp}
\title{Function that assign a window from a list to the correct ppp in another list}
\usage{
win2ppp(X, W)
}
\arguments{
\item{X}{list (solist) of ppp objects to which change the Window.}

\item{W}{list (solist) of Window objects}
}
\description{
The function match the names of the objects in the two lists to make sure that the correct Window is assigned to the corresponding ppp.
}
\details{
It should work either when the windows list is as long or longer than the ppp list, or vice versa.
}
\seealso{
\link{solist}, \link{Window}
}
\author{
Filippo Ferrario, \email{filippo.f3rrario@gmail.com}
}
