% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoWin.R
\name{MoWin}
\alias{MoWin}
\title{Variable squared moving window}
\usage{
MoWin(dist = NULL, side.cell = NULL)
}
\arguments{
\item{dist}{width of the square}

\item{side.cell}{width of the cell/pixel of the moving window}
}
\description{
Prepare a squared moving window matrix of variable size to be used in other functions (e.g., raster::focal, landscapemetrics::window_lsm)
}
