% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_patches.R
\name{PatchNum}
\alias{PatchNum}
\title{Number of patch in buffer}
\usage{
PatchNum(poly, buffer, cx)
}
\arguments{
\item{poly}{a collection of polygons representing the patches to be use to calculate the index}

\item{buffer}{a vector (or dataframe) of polygons representing the buffers. tipically the result of sf::st_buffer.}

\item{cx}{st_point center(s) of the buffer. Provided by \code{map2raster} if used in it.}
}
\description{
Number of patch in buffer
}
\seealso{
\link{map2raster}, \link{PatchRelv}, \link[sf:st]{sf::st_point}, \link[sf:geos_unary]{sf::st_buffer}
}
\author{
Filippo Ferrario, \email{filippo.f3rrario@gmail.com}
}
